+++ tests/categories_test.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
-from collections import defaultdict
 import unittest
 import wikipediaapi
 
-from mock_data import wikipedia_api_request
+from tests.mock_data import wikipedia_api_request
 
 
 class TestCategories(unittest.TestCase):
@@ -28,3 +27,7 @@ class TestCategories(unittest.TestCase):
             list(sorted(map(lambda s: s.ns, page.categories.values()))),
             [14] * 3
         )
+
+    def test_no_categories_count(self):
+        page = self.wiki.page('No_Categories')
+        self.assertEqual(len(page.categories), 0)