+++ tests/extract_errors_test.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
-from collections import defaultdict
 import unittest
 
-from mock_data import wikipedia_api_request
+from tests.mock_data import wikipedia_api_request
 import wikipediaapi